move api logic from /services into /api directory
