+++ src/routes/_auth/generate.$orderId.tsx
@@ -1,6 +1,8 @@
 import { createFileRoute, useParams } from "@tanstack/react-router";
 import { Image, Pullover } from "abipulli-types";
 import { useEffect, useState } from "react";
+import { ImageApi } from "src/api/endpoints/image";
+import { PulloverApi } from "src/api/endpoints/pullover";
 import { BackButton } from "src/components/Buttons/BackButton";
 import { SuperDuperShinySpecialButton } from "src/components/Buttons/SuperDuperShinySpecialButton";
 import { ExampleSection } from "src/components/Generate/ExampleSection";
@@ -12,8 +14,6 @@ import { InputField } from "src/components/Inputs/InputField";
 import { SmallLabel } from "src/components/Texts/SmallLabel";
 import { GENERATION_EXAMPLES } from "src/exampleData/GenerationExamples";
 import { useSnackbar } from "src/hooks/useSnackbar";
-import { ImageApi } from "src/services/endpoints/image";
-import { PulloverApi } from "src/services/endpoints/pullover";
 import { GenerationExample } from "src/types/generator/GenerationExample";
 import { SelectedStylesMap } from "src/types/generator/SelectedStyles";
